feat(api): move #createOrganizationsWithTagsAndTargetProfiles usecase to right context
